video-vector/requirements.txt
fastapi
uvicorn[standard]
python-multipart 
# 深度学习框架
torch==2.3.1
torchvision==0.18.1
torchaudio
# 图像和文本处理
transformers
# 视频和音频处理
opencv-python==4.9.0.80
librosa
funasr>=1.1.3
numpy<=1.26.4

modelscope
huggingface
huggingface_hub
Pillow
httpx
einops
timm

sympy==1.13.1